Add utils and validation
